pyproject.toml
[project]
name = "hardrockeresportsdatabasebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
]
